Milestone-VP Admin-Dashboard-Management
Implemented Create, Edit, and Delete functionalities for parking lots.
Implemented functionalities to View details of parking lots and status of parking spots. Further, incorporated functionalities to Add and Delete a parking spot.
Implemented automatic creation of parking spots based on the number of parking spots in the parking lot.
Created users tab to display the details of all the users.